Paramter base get request added for overall function, no date and time added

+++ BoxStoreApp/views.py
@@ -5,8 +5,14 @@ from django.views.decorators.csrf import csrf_exempt
 from django.contrib.auth import authenticate, login, logout, get_user_model
 User = get_user_model()
 from .models import BoxModel
-
-
+from rest_framework.authtoken.models import Token
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.permissions import IsAuthenticated, IsAdminUser
+from rest_framework.decorators import authentication_classes, permission_classes
+from django.db.models import Sum, Q
+from django.conf import settings
+from datetime import datetime
+# settings.configure()
 
 
 def hello(req):
@@ -24,16 +30,19 @@ def is_staff(req):
     else:
         return False
 
-
-
 @csrf_exempt
 def register(req):
     if req.method == 'POST':
         username = req.POST['username']
         password = req.POST['password']
         usertype = req.POST['type']
-        user = User.objects.create_user(username, usertype, password)
+        if usertype == 'staff':
+            usertype = True
+        else:
+            usertype = False
+        user = User.objects.create_user(username= username, is_staff= usertype, password=password)
         user.save()
+
         return HttpResponse('User created')
     else:
         return HttpResponse('Error creating user')
@@ -45,7 +54,9 @@ def loginx(req):
         user = authenticate(req, username=username, password=password)
         if user is not None:
             login(req, user)
-            return HttpResponse('User logged in'+ str(user.username)+': ' +str(user.typex)+ ':'+ str(user.is_staff))
+            token, created= Token.objects.get_or_create(user=user)
+            print(token.key)
+            return JsonResponse({'status': 200, 'token': token.key, 'user': user.username, 'isStaff': user.is_staff})
         else:
             return HttpResponse('Error logging in')
     else:
@@ -63,9 +74,12 @@ def isLogged(req):
     else:
         return HttpResponse('User is not logged in')
 
+
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 @csrf_exempt
 def addBox(req):
-    if req.method == 'POST':
+    if req.method == 'POST' and is_staff(req):
         boxName = req.POST['name']
         boxLength = int(req.POST['length'])
         boxWidth = int(req.POST['width'])
@@ -81,13 +95,15 @@ def addBox(req):
         print(BoxModel.objects.all().values())
         return HttpResponse('Box created')
     else:
-        return HttpResponse('Error creating box')
+        return HttpResponse('Error creating box'+ str(req.user.is_authenticated)+ str(req.user.is_staff))
 
 
 # Create your views here.
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 @csrf_exempt
 def updateBox(req, id):
-    if req.method == 'POST':
+    if req.method == 'POST' and is_staff(req):
         boxName = req.POST['name']
         boxLength = int(req.POST['length'])
         boxWidth = int(req.POST['width'])
@@ -109,13 +125,68 @@ def updateBox(req, id):
     else:
         return HttpResponse('Error updating box')
 
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 @csrf_exempt
 def deleteBox(req, id):
     # if req.method == 'DELETE' and is_authenticated(req) and is_staff(req):
-    if req.method == 'DELETE':
-
+    if req.method == 'DELETE' and is_staff(req):
         box = BoxModel.objects.get(id=id)
         box.delete()
         return HttpResponse('Box deleted')
     else:
         return HttpResponse('Error deleting box')
+
+@csrf_exempt
+def listBox(req):
+    if req.method == 'GET' and is_authenticated(req):
+        boxes = BoxModel.objects.all().values()
+        total_area= BoxModel.objects.aggregate(Sum('area'))
+        total_volume= BoxModel.objects.aggregate(Sum('volume'))
+        total_individual_box_count= BoxModel.objects.filter(created_by=req.user.username).count()
+        total_boxes= BoxModel.objects.all().count()
+        filtered_data= filter_boxes(req)
+        return JsonResponse({'status': 200, 'boxes': list(boxes), 'filtered Data': str(filtered_data), 'user': str(req.user), 'total_area': total_area, 'total_volume': total_volume, 'total_individual_area': total_individual_box_count, 'total_boxes': total_boxes})
+    else:
+        return HttpResponse('Error listing boxes')
+
+
+
+
+def filter_boxes(request):
+    if request.method== 'GET':
+        length_more_than = request.GET.get('length_more_than', 0)
+        area_more_than = request.GET.get('area_more_than', 0)
+        volume_more_than = request.GET.get('volume_more_than', 0)
+        height_more_than = request.GET.get('height_more_than', 0)
+        width_more_than = request.GET.get('width_more_than', 0)
+
+        # just for a maximum value as float('inf') is not working
+        maxm= 10000000000
+        length_less_than = request.GET.get('length_less_than', maxm)
+        area_less_than = request.GET.get('area_less_than', maxm)
+        volume_less_than = request.GET.get('volume_less_than', maxm)
+        height_less_than = request.GET.get('height_less_than', maxm)
+        width_less_than = request.GET.get('width_less_than', maxm)
+
+        created_by = request.GET.get('created_by', None)
+
+        # print('ll:', length_more_than, length_less_than)
+        boxes= BoxModel.objects.filter(
+            Q(length__gt=length_more_than) & 
+            Q(area__gt=area_more_than) & 
+            Q(volume__gt=volume_more_than) & 
+            Q(height__gt=height_more_than) & 
+            Q(width__gt=width_more_than)&
+            Q(length__lt= length_less_than)&
+            Q(width__lt= width_less_than)&
+            Q(height__lt= height_less_than)&
+            Q(area__lt= area_less_than)&
+            Q(volume__lt= volume_less_than)
+            )
+        
+        if created_by:
+            boxes= boxes.filter(created_by= created_by)
+        
+        # print(boxes.values())
+        return boxes.values()
